Add files via upload
NOTE FOR THE READER /EVALUATOR:-

VALIDATION:-     form validation is done through the HTML FORM VALIDATION technique , for eg: - for email attribute type is set to "email" value so that proper validation is done  also for each field "REQUIRED " attribute value is being set so that none of the fields is left empty.  
UPLOADS :-     3 files ( index2  , table :- which shows the data , and insert)  along with 3 screenshots are uploaded for the code. Database name is " registered_students" and the table name is " students".
CSS :- Bootstrap is uesd for the styling purposes alongwith inline css whrerever required.
THANK YOU !